fix: update for Haste 1.6.a
Haste 1.6.a reimplements the whole networking system. Primarily, it reimplements HasteNetworking.SetState to be a bit more robust, and to always setup steam networking automatically.
As you can imagine, this is bad for us in the no-Steam networking land.

The "fix" here is to do the same things that `HasteNetworking.SetState()` does, but setting up our own networking transport instead.

Overall, its just a bunch of refactoring on our side too.